python/requirements.txt
Flask
yt-dlp
pydub
google-cloud-speech
openai
google.generativeai